move requests service work behind a toggle
